# packages/contratos/openapi.json
{
  "components" : {
    "schemas" : {
...
      "CrearIntentoDePagoRequest" : {
        "properties" : {
          "pedidoId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "DineroRespuesta" : {
        "properties" : {
          "moneda" : {
            "type" : "string"
          },
          "valor" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [
          "moneda",
          "valor"
        ],
        "type" : "object"
      },
...
      "IntentoDePagoRespuesta" : {
        "properties" : {
          "ambiente" : {
            "type" : "string"
          },
          "firmaIntegridad" : {
            "type" : "string"
          },
          "llavePublica" : {
            "type" : "string"
          },
          "monto" : {
            "$ref" : "#/components/schemas/DineroRespuesta"
          },
          "montoEnCentavos" : {
            "format" : "int64",
            "type" : "integer"
          },
          "referencia" : {
            "type" : "string"
          }
        },
        "required" : [
          "montoEnCentavos"
        ],
        "type" : "object"
      },
...
      "RegistrarIdTransaccionWompiRequest" : {
        "properties" : {
          "idTransaccionWompi" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/envios/webhook" : {
      "post" : {
        "operationId" : "webhook_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "format" : "byte",
                "type" : "string"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [
          "envio-webhook-controlador"
        ]
      }
    },
...
    "/api/v1/pagos/intentos" : {
      "post" : {
        "operationId" : "crear_2",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CrearIntentoDePagoRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IntentoDePagoRespuesta"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [
          "pago-controlador"
        ]
      }
    },
    "/api/v1/pagos/intentos/{referencia}" : {
      "patch" : {
        "operationId" : "registrarIdTransaccion",
        "parameters" : [
          {
            "in" : "path",
            "name" : "referencia",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegistrarIdTransaccionWompiRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [
          "pago-controlador"
        ]
      }
    },
...
    "/api/v1/pagos/webhook" : {
      "post" : {
        "operationId" : "webhook",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "additionalProperties" : true,
                "type" : "object"
              }
            }
          },
          "description" : "Evento de Wompi, tal como lo envía la pasarela. Se valida por firma.",
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [
          "pago-controlador"
        ]
      }
    },
...
  }
}